iot_devices/requirements.txt
requests==2.28.2
paho-mqtt==1.6.1